+++ src/main/java/org/opentripplanner/graph_builder/module/osm/OpenStreetMapModule.java
@@ -291,8 +291,10 @@ public class OpenStreetMapModule implements GraphBuilderModule {
                 if (operators != null)
                     networkSet.addAll(Arrays.asList(operators.split(";")));
                 if (networkSet.isEmpty()) {
-                    LOG.warn("Bike rental station at osm node " + node.getId() + " ("
-                            + creativeName + ") with no network; including as compatible-with-all.");
+                    LOG.warn(graph
+                            .addBuilderAnnotation(new BikeRentalStationNoNetwork(node.getId(),
+                                    creativeName, node.lon, node.lat))
+                            + " Including as compatible-with-all.");
                     networkSet = null; // Special "catch-all" value
                 }
                 BikeRentalStation station = new BikeRentalStation();
@@ -467,6 +469,9 @@ public class OpenStreetMapModule implements GraphBuilderModule {
                     }
                 }
             }
+            // Place the P+R at the center of the envelope
+            double centerX = (envelope.getMinX() + envelope.getMaxX()) / 2;
+            double centerY = (envelope.getMinY() + envelope.getMaxY()) / 2;
             // Check P+R accessibility by walking and driving.
             TraversalRequirements walkReq = new TraversalRequirements(new RoutingRequest(
                     TraverseMode.WALK));
@@ -499,7 +504,9 @@ public class OpenStreetMapModule implements GraphBuilderModule {
             }
             if (!walkAccessibleOut || !carAccessibleIn) {
                 // This will prevent the P+R to be useful.
-                LOG.warn(graph.addBuilderAnnotation(new ParkAndRideUnlinked((creativeName != null ? creativeName.toString() : "null"), osmId)));
+                LOG.warn(graph.addBuilderAnnotation(new ParkAndRideUnlinked(
+                        (creativeName != null ? creativeName.toString() : "null"), osmId, centerX,
+                        centerY)));
                 return false;
             }
             if (!walkAccessibleIn || !carAccessibleOut) {
@@ -889,9 +896,9 @@ public class OpenStreetMapModule implements GraphBuilderModule {
                         }
                         for (StreetEdge to : restrictionTag.possibleTo) {
                             if (from == null || to == null) {
-                                graph.addBuilderAnnotation(
-                                    new TurnRestrictionBad(restrictionTag.relationOSMID,
-                                        "to-edge is null"));
+                                graph.addBuilderAnnotation(new TurnRestrictionBad(
+                                        restrictionTag.relationOSMID, "to-edge is null",
+                                        from.getToVertex().getLon(), from.getToVertex().getLat()));
                                 continue;
                             }
                             int angleDiff = from.getOutAngle() - to.getInAngle();
@@ -902,32 +909,40 @@ public class OpenStreetMapModule implements GraphBuilderModule {
                             case LEFT:
                                 if (angleDiff >= 160) {
                                     graph.addBuilderAnnotation(
-                                        new TurnRestrictionBad(restrictionTag.relationOSMID,
-                                            "Left turn restriction is not on edges which turn left"));
+                                            new TurnRestrictionBad(restrictionTag.relationOSMID,
+                                                    "Left turn restriction is not on edges which turn left",
+                                                    from.getToVertex().getLon(),
+                                                    from.getToVertex().getLat()));
                                     continue; // not a left turn
                                 }
                                 break;
                             case RIGHT:
                                 if (angleDiff <= 200) {
                                     graph.addBuilderAnnotation(
-                                        new TurnRestrictionBad(restrictionTag.relationOSMID,
-                                            "Right turn restriction is not on edges which turn right"));
+                                            new TurnRestrictionBad(restrictionTag.relationOSMID,
+                                                    "Right turn restriction is not on edges which turn right",
+                                                    from.getToVertex().getLon(),
+                                                    from.getToVertex().getLat()));
                                     continue; // not a right turn
                                 }
                                 break;
                             case U:
                                 if ((angleDiff <= 150 || angleDiff > 210)) {
                                     graph.addBuilderAnnotation(
-                                        new TurnRestrictionBad(restrictionTag.relationOSMID,
-                                            "U-turn restriction is not on U-turn"));
+                                            new TurnRestrictionBad(restrictionTag.relationOSMID,
+                                                    "U-turn restriction is not on U-turn",
+                                                    from.getToVertex().getLon(),
+                                                    from.getToVertex().getLat()));
                                     continue; // not a U turn
                                 }
                                 break;
                             case STRAIGHT:
                                 if (angleDiff >= 30 && angleDiff < 330) {
                                     graph.addBuilderAnnotation(
-                                        new TurnRestrictionBad(restrictionTag.relationOSMID,
-                                            "Straight turn restriction is not on edges which go straight"));
+                                            new TurnRestrictionBad(restrictionTag.relationOSMID,
+                                                    "Straight turn restriction is not on edges which go straight",
+                                                    from.getToVertex().getLon(),
+                                                    from.getToVertex().getLat()));
                                     continue; // not straight
                                 }
                                 break;